server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
transformers>=4.45.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
google-genai>=1.0.0
motor>=3.3.0
beanie>=1.25.0
python-dotenv>=1.0.0
